fixed bug when new private key was created after skipping domain

+++ init-letsencrypt.sh
@@ -1,3 +1,4 @@
+
 #!/bin/bash
 
 domains=("example.com www.example.com" "another-example.com www.another-example.com")
@@ -50,10 +51,13 @@ for domain in ${!domains[*]}; do
   
   mkdir -p "$data_path/conf/live/$domain_name"
 
-  echo "### Creating dummy certificate for $domain_name domain..."
-  path="/etc/letsencrypt/live/$domain_name"
-  docker-compose run --rm --entrypoint "openssl req -x509 -nodes -newkey rsa:1024 \
-  -days 1 -keyout '$path/privkey.pem' -out '$path/fullchain.pem' -subj '/CN=localhost'" certbot
+  if [ ! -e "$data_path/conf/live/$domain_name/cert.pem" ]; then
+    echo "### Creating dummy certificate for $domain_name domain..."
+    path="/etc/letsencrypt/live/$domain_name"
+    docker-compose run --rm --entrypoint "openssl req -x509 -nodes -newkey rsa:1024 \
+    -days 1 -keyout '$path/privkey.pem' -out '$path/fullchain.pem' -subj '/CN=localhost'" certbot
+  fi
+
 done
 
 echo "### Starting nginx ..."
